+++ nncf/quantization/algorithms/bias_correction/backend.py
@@ -22,7 +22,7 @@ from nncf.common.graph.transformations.commands import TransformationCommand
 from nncf.common.tensor import NNCFTensor
 from nncf.common.graph import NNCFGraph
 from nncf.common.graph import NNCFNode
-from nncf.common.tensor_statistics.collectors import TensorReducerBase
+from nncf.common.tensor_statistics.collectors import TensorStatisticCollectorBase
 from nncf.common.tensor_statistics.collectors import ReductionShape
 from nncf.common.utils.registry import Registry
 
@@ -110,7 +110,7 @@ class BiasCorrectionAlgoBackend(ABC):
     @abstractmethod
     def mean_statistic_collector(reduction_shape: ReductionShape,
                                  num_samples: Optional[int] = None,
-                                 window_size: Optional[int] = None) -> TensorReducerBase:
+                                 window_size: Optional[int] = None) -> TensorStatisticCollectorBase:
         """
         Returns backend-specific mean statistic collector.
 
@@ -122,7 +122,7 @@ class BiasCorrectionAlgoBackend(ABC):
 
     @staticmethod
     @abstractmethod
-    def batch_statistic_collector(num_samples: int = None) -> TensorReducerBase:
+    def batch_statistic_collector(num_samples: int = None) -> TensorStatisticCollectorBase:
         """
         Returns backend-specific batch statistic collector.
