+++ tests/post_training/model_scope.py
@@ -58,7 +58,7 @@ def get_validation_scope() -> List[dict]:
     model_scope.append({"name": "dpn68", "quantization_params": {"preset":QuantizationPreset.MIXED}})
     model_scope.append({"name": "dla34", "quantization_params": {"preset":QuantizationPreset.MIXED}})
 
-    return model_scope[0:1]
+    return model_scope
 
 
 VALIDATION_SCOPE = get_validation_scope()
