+++ tests/post_training/requirements.txt
@@ -8,4 +8,3 @@ pytest
 openvino-dev>=2022.1
 timm
 sklearn
-pandas
