+++ tests/tensorflow/tensor_statistics/test_tensor_statistics.py
@@ -18,7 +18,7 @@ import pytest
 import tensorflow as tf
 
 from nncf.common.tensor_statistics.statistics import TensorStatistic
-from nncf.common.tensor_statistics.collectors import TensorReducerBase, ReductionShape, \
+from nncf.common.tensor_statistics.collectors import TensorStatisticCollectorBase, ReductionShape, \
     StatisticsNotCollectedError, OfflineTensorStatisticCollector
 from nncf.tensorflow.tensor_statistics.collectors import TFMinMaxStatisticCollector
 from nncf.tensorflow.tensor_statistics.collectors import TFMedianMADStatisticCollector
@@ -104,7 +104,7 @@ class TestCollectedStatistics:
                                          }
                                  )
                              ])
-    def test_collected_statistics_with_shape_convert(self, collector: Type[TensorReducerBase],
+    def test_collected_statistics_with_shape_convert(self, collector: Type[TensorStatisticCollectorBase],
                                                      reduction_shapes_vs_ref_statistic:
                                                      Dict[Tuple[ReductionShape, ReductionShape], TensorStatistic]):
         for reduction_shape in reduction_shapes_vs_ref_statistic.keys():
@@ -186,7 +186,7 @@ class TestCollectedStatistics:
                                          }
                                  ),
                              ])
-    def test_collected_statistics(self, collector: Type[TensorReducerBase],
+    def test_collected_statistics(self, collector: Type[TensorStatisticCollectorBase],
                                   reduction_shapes_vs_ref_statistic: Dict[ReductionShape, TensorStatistic]):
         for reduction_shape in reduction_shapes_vs_ref_statistic.keys():
             collector_obj = collector(reduction_shape=reduction_shape)
@@ -215,12 +215,12 @@ class TestCollectedStatistics:
         collector_type = request.param
         return collector_type(reduction_shape=(1,))
 
-    def test_collected_samples(self, collector_for_interface_test: TensorReducerBase):
+    def test_collected_samples(self, collector_for_interface_test: TensorStatisticCollectorBase):
         for input_ in TestCollectedStatistics.REF_INPUTS:
             collector_for_interface_test.register_input(input_)
         assert collector_for_interface_test.collected_samples() == len(TestCollectedStatistics.REF_INPUTS)
 
-    def test_reset(self, collector_for_interface_test: TensorReducerBase):
+    def test_reset(self, collector_for_interface_test: TensorStatisticCollectorBase):
         for input_ in TestCollectedStatistics.REF_INPUTS:
             collector_for_interface_test.register_input(input_)
         collector_for_interface_test.reset()
@@ -228,7 +228,7 @@ class TestCollectedStatistics:
         with pytest.raises(StatisticsNotCollectedError):
             collector_for_interface_test.get_statistics()
 
-    def test_enable_disable(self, collector_for_interface_test: TensorReducerBase):
+    def test_enable_disable(self, collector_for_interface_test: TensorStatisticCollectorBase):
         for input_ in TestCollectedStatistics.REF_INPUTS:
             collector_for_interface_test.register_input(input_)
